Complete the Lwjgl3 IO backend

+++ src/main/java/com/shc/silenceengine/backend/lwjgl3/io/Lwjgl3DirectBuffer.java
@@ -78,7 +78,7 @@ public class Lwjgl3DirectBuffer extends DirectBuffer
     }
 
     @Override
-    public DirectBuffer writeInt(int byteIndex, byte value)
+    public DirectBuffer writeByte(int byteIndex, byte value)
     {
         nativeBuffer.put(byteIndex, value);
         return this;
